% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.between_df.R
\name{f.between_df}
\alias{f.between_df}
\title{Function to compute between standard deviation (based on id)}
\usage{
f.between_df(data, id, col, na.rm = TRUE)
}
\arguments{
\item{data}{a dataframe}

\item{id}{a grouping variable}

\item{col}{column names (character)}

\item{na.rm}{a logical. Should missing values be removed?}
}
\value{
A dataframe with the between standard deviation of each col variables
}
\description{
Function to compute between standard deviation (based on id)
}
\examples{
f.between_df(iris, Species, c("Sepal.Width", "Sepal.Length"))

}
