% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.overall_df.R
\name{f.overall_df}
\alias{f.overall_df}
\title{Function to compute overall standard deviation}
\usage{
f.overall_df(data, id, col, na.rm = TRUE)
}
\arguments{
\item{data}{a dataframe}

\item{id}{a grouping variable}

\item{col}{column names (character)}

\item{na.rm}{a logical. Should missing values be removed?}
}
\value{
A dataframe with the overall standard deviation of each col
}
\description{
This function allows to compute the overall standard deviation of a variable and returns a nice table.
}
\examples{
f.overall_df(iris, Species, c("Sepal.Width", "Sepal.Length"))
}
