% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.within_df.R
\name{f.within_df}
\alias{f.within_df}
\title{Function to compute within standard deviation (based on id)}
\usage{
f.within_df(data, id, col, na.rm = TRUE)
}
\arguments{
\item{data}{a dataframe}

\item{id}{a grouping variable}

\item{col}{column names (character)}

\item{na.rm}{a logical. Should missing values be removed?}
}
\value{
A dataframe with the within sd of each col variables
}
\description{
This function allows to compute the within standard deviation of any grouped numeric variable
}
\examples{
f.within_df(iris, Species, c("Sepal.Width", "Sepal.Length"))
}
