% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rxtsum.R
\name{Rxtsum}
\alias{Rxtsum}
\title{Rxtsum
Rxtsum}
\usage{
Rxtsum(data, id, col, na.rm = TRUE)
}
\arguments{
\item{data}{a dataframe}

\item{id}{a grouping variable}

\item{col}{column names (character)}

\item{na.rm}{a logical. Should missing values be removed?}
}
\value{
A dataframe with between/within/overall standard deviation for each col variables defined.
}
\description{
This function allows to compute the between/within standard deviation of any grouped variables. It replicates Stata's xtsum methodology and returns a dataframe which is convenient for publication with package like kable.
}
\examples{
Rxtsum(iris, Species, c("Sepal.Width", "Sepal.Length"))
}
